studentAttendance/requirements.txt
absl-py==1.3.0
asgiref==3.6.0
astunparse==1.6.3
autopep8==2.0.2
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Cython==0.29.34
Django==4.2
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.2.0
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
nltk==3.8
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pillow==10.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycodestyle==2.10.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SpeechRecognition==3.9.0
sqlparse==0.4.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.3.0.dev221212
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow_intel==2.10.0
termcolor==2.1.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
zipp==3.11.0
